% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_brightness.R
\name{adjust_brightness}
\alias{adjust_brightness}
\alias{darken}
\alias{lighten}
\title{Adjust the brightness of a color}
\usage{
adjust_brightness(color, amount)

lighten(color, amount)

darken(color, amount)
}
\arguments{
\item{color}{The color}

\item{amount}{A value between -1 and 1.}
}
\value{
The adjusted color as a hexadecimal string
}
\description{
\code{adjust_brightness} increases the brightness of a color by
the amount given. Resulting brightness is clamped to the range [0,1].

\code{lighten} increases the brightness of the color by the
given amount

\code{darken} decreases the brightness of the color by the
given amount
}
\note{
Adjustments are made in the HSV colorspace
}
\examples{
lighten('orange', 0.2)

}

