% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hsl.R
\name{adjust_hsl}
\alias{adjust_hls}
\alias{adjust_hsl}
\title{Adjust color properties in HSL space}
\usage{
adjust_hsl(color, Hamount = 0, Samount = 0, Lamount = 0, wraphue = TRUE)
}
\arguments{
\item{color}{One or more colors, either hex or named}

\item{Hamount}{Amount to adjust hue (default: 0)}

\item{Samount}{Amount to adjust saturation (default: 0)}

\item{Lamount}{Amount to adjust lightness (default: 0)}

\item{wraphue}{Whether or not the resulting hue should be constrained to
[0,360) (default: TRUE)}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
\code{adjust_hsl} (also \code{adjust_hls}) adjusts the
properties of colors in HSL space(hue, saturation, and lightness)
}
\note{
Adjusted saturation and brightness are clamped to the range [0,1].

If \code{wraphue} is \code{FALSE}, resulting hue is clamped to the
range [0,360].
}
\examples{
# Decrease saturation of orange by 0.1
adjust_hsl(color = "orange", Samount = -0.1)


}

