% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hsv.R
\name{adjust_hsv}
\alias{adjust_hsv}
\title{Adjust color properties in HSV space}
\usage{
adjust_hsv(col, Hamount = 0, Samount = 0, Vamount = 0)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{Hamount}{Amount to adjust hue (default: \code{0})}

\item{Samount}{Amount to adjust saturation (default: \code{0})}

\item{Vamount}{Amount to adjust value/brightness (default: \code{0})}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
\code{adjust_hsv} adjusts the properties of colors in HSV space
(hue, saturation, and value/brightness)
}
\note{
Adjusted saturation and brightness are clamped to the range [0,1].

If \code{wraphue} is \code{FALSE}, resulting hue is clamped to the
range [0,360].
}
\examples{
# Decrease saturation of orange by 0.1
adjust_hsv(color = "orange", Samount = -0.1)

lighten('orange', 0.2)

}

