% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hue.R
\name{adjust_hue}
\alias{adjust_hue}
\alias{complement}
\title{Adjust a color's hue}
\usage{
adjust_hue(color, amount, wrap = TRUE, space = "HSV")

complement(color, wrap = TRUE, space = "HSV")
}
\arguments{
\item{color}{The color}

\item{amount}{The amount, in degrees}

\item{wrap}{Whether or not the resulting hue should be constrained to [0,360)}

\item{space}{The colorspace, HSV (default), HSL, or HLS}
}
\value{
The adjusted color as a hexadecimal string
}
\description{
\code{adjust_hue} adjusts a color's hue by a given amount in HSV
space. The \code{wrap} parameter, enabled by default, allows the hue value to
"wrap around", keeping hue within the range [0,360).

\code{complement} returns the complement of the color by
adjusting the color's hue by 180 degrees in HSV space
}
\examples{
adjust_hue("#807040", amount = 45)
complement("#607080")

}

