% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rgb.R
\name{adjust_rgb}
\alias{adjust_rgb}
\title{Adjust a color's RGB components}
\usage{
adjust_rgb(col, r = NULL, g = NULL, b = NULL, mode = "relative")
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{r}{The adjustment in hue [-1,1] (default: \code{NULL})}

\item{g}{The adjustment in saturation [-1,1] (default: \code{NULL})}

\item{b}{The adjustment in brightness [-1,1] (default: \code{NULL})}

\item{mode}{Whether the adjustments set the value directly
("\code{absolute}") or are added to the current value ("\code{relative}",
default).}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
These functions adjust a color's red, green, and blue components

\code{adjust_rgb} adjusts red, green, and blue, either
by replacing the current value (\code{mode="absolute"}) or adding to it
(\code{mode="relative"}, the default).
}
\examples{
# Add red to purple
adjust_rgb("purple", r = 0.2)

}
\seealso{
\code{\link{red}}, \code{\link{green}}, \code{\link{blue}}, and
\code{\link{alpha}} to get RGB component values
}

