% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hsv.R
\name{col2hsv}
\alias{col2hsv}
\title{Color to HSV Conversion}
\usage{
col2hsv(col)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}
}
\value{
A matrix with a column for each color. The three rows of the matrix
indicate hue ("\code{h}"), saturation ("\code{s}"), and value ("\code{v}"),
which is also referred to as \emph{brightness}).
}
\description{
Convert \R colors to HSV (hue/saturation/value)
}
\note{
The alpha channel is stripped from hexadecimal strings where one is
included
}
\examples{
# Get HSV coordinates for red
col2hsv("red")

# Get HSV coordinates for several different colors
col2hsv(c("#FF0000", "mediumpurple3", 2, "steelblue4"))

}

