% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colHLS.R
\name{colHLS}
\alias{colHLS}
\title{Convert a hex or named color to a colorspace HLS object}
\usage{
colHLS(color)
}
\arguments{
\item{color}{The color}
}
\value{
An object of class RGB.
}
\description{
Convert a hex or named color to a colorspace HLS object
}
\note{
Alpha values are ignored
}
\examples{
colHLS("orange")

}
\seealso{
\code{\link[colorspace]{HLS}}
}

