% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colHSV.R
\name{colHSV}
\alias{colHSV}
\title{Convert a hex or named color to a colorspace HSV object}
\usage{
colHSV(color)
}
\arguments{
\item{color}{The color}
}
\value{
An object of class RGB.
}
\description{
Convert a hex or named color to a colorspace HSV object
}
\note{
Alpha values are ignored
}
\examples{
colHSV("orange")

}
\seealso{
\code{\link[colorspace]{HSV}}
}

