% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorproperty.R
\name{colorproperty}
\alias{blue}
\alias{colorproperty}
\alias{green}
\alias{red}
\title{Get values of color properties in different color spaces}
\usage{
colorproperty(color, space, property)

red(color)

green(color)

blue(color)
}
\arguments{
\item{color}{One or more colors, either hex or named}

\item{space}{The color space}

\item{property}{The property value to get. For example, when \code{space} is
\code{HSV}, \code{property} can be "H", "S", or "V".}
}
\value{
Numeric property values
}
\description{
Get values of color properties in different color spaces
}
\examples{
colorproperty(color = "mediumpurple3", space = "HSV", property = "S")

}

