% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colRGB.R
\name{colRGB}
\alias{colRGB}
\title{Convert a hex or named color to a colorspace RGB object}
\usage{
colRGB(color)
}
\arguments{
\item{color}{The color}
}
\value{
An object of class RGB.
}
\description{
Convert a hex or named color to a colorspace RGB object
}
\note{
Alpha values are ignored
}
\examples{
colRGB("orange")

}
\seealso{
\code{\link[colorspace]{RGB}}
}

