% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement}
\alias{complement}
\title{Return the complement of the color (adjust hue by 180 degrees)}
\usage{
complement(color, space = "HSV", wrap = TRUE)
}
\arguments{
\item{color}{One or more colors, either hex or named}

\item{space}{The colorspace to use, "\code{HSV}" (default) or "\code{HSL}"}

\item{wrap}{Whether or not the resulting hue should be constrained to
[0,360] (default: \code{TRUE})}
}
\value{
The adjusted colors as hexadecimal strings
}
\description{
Return the complement of the color (adjust hue by 180 degrees)
}
\examples{
complement("orange")

}

