% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\title{Get color coordinates in a given color space}
\usage{
coordinates(color, space = "sRGB")
}
\arguments{
\item{color}{One or more colors, either hex or named}

\item{space}{The colorspace to use}
}
\value{
A \code{\link[colorspace]{color-class}} object
}
\description{
Get color coordinates in a given color space
}
\examples{
coordinates(color = "orange")

}

