% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex}
\alias{hex}
\title{Get a hex string representation of colors}
\usage{
hex(color)
}
\arguments{
\item{color}{One or more colors, either hex or named}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
Get a hex string representation of colors
}
\examples{
hex("orange")

}

