% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv2col.R
\name{hsv2col}
\alias{hsv2col}
\title{HSV to R color conversion}
\usage{
hsv2col(col)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}
}
\value{
A vector of colors (as hexadecimal strings) corresponding to the
given values in HSV space
}
\description{
\code{hsv2col} transforms colors from HSV space into   TODO
}
\examples{
TODO

}

