% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv2hex.R
\name{hsv2hex}
\alias{hsv2hex}
\title{HSV to Hex Color Conversion}
\usage{
hsv2hex(col, alpha = FALSE)
}
\arguments{
\item{col}{A matrix containing three rows (\code{h}, \code{s}, \code{v}) and
a column for each color specified. Values are in [0,1].}

\item{alpha}{Whether or not to include alpha channel values in the result.
Because HSV matrices do not have this information, full opacity ("FF") is
used.}
}
\value{
A vector of colors (as hexadecimal strings) corresponding to the
given values in HSV space
}
\description{
Convert matrices containing "hue", "saturation", and "value"
values into hex color strings.
}
\examples{
hsv2hex(col2hsv("orange"))

}

