% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv_properties.R
\name{hue}
\alias{brightness}
\alias{hue}
\alias{saturation}
\title{View a color's HSV components}
\usage{
hue(col)

saturation(col)

brightness(col)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}
}
\value{
A number indicating the component value [0,1]
}
\description{
\code{hue}, \code{saturation}, and \code{brightness} return the values for
those components
}
\examples{
# Get the saturation of purple
saturation("purple")
}
\seealso{
\code{\link{adjust_hsv}} to manipulate these components
}

