% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hue.R
\name{hue}
\alias{complement}
\alias{hue}
\title{View or adjust a color's hue}
\usage{
hue(color, space = "HSV")

complement(color, space = "HSV", wrap = TRUE)
}
\arguments{
\item{color}{One or more colors, either hex or named}

\item{space}{The colorspace to use, "\code{HSV}" (default) or "\code{HSL}"}

\item{wrap}{Whether or not the resulting hue should be constrained to
[0,360] (default: \code{TRUE})}
}
\value{
The adjusted colors as hexadecimal strings
}
\description{
\code{hue} returns the color's hue [0,360)

\code{complement} returns a color's complement, which is the hue
adjusted by 180 degrees
}
\examples{
hue("orange")

complement("orange", space = "HSL")

}

