% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert}
\alias{invert}
\title{Invert a color}
\usage{
invert(color)
}
\arguments{
\item{color}{The color to invert}
}
\value{
The inverted color as a hexadecimal string
}
\description{
\code{invert} inverts each channel of a color
}
\examples{
invert('#FF00FF')

}

