% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is.color}
\alias{is.color}
\alias{is.hex.color}
\alias{is.named.color}
\title{Check whether input is a valid color}
\usage{
is.color(x, allow_alpha = TRUE)

is.hex.color(x, allow_alpha = TRUE)

is.named.color(x)
}
\arguments{
\item{x}{A color string}

\item{allow_alpha}{Whether or not alpha parameters are allowed in hex color
strings (default: \code{TRUE})}
}
\value{
Boolean indicating whether or not the given input is a valid color
}
\description{
\code{is.color} checks whether or not the given input is a valid
color

\code{is.hex.color} checks whether input is a valid hex color
(with or without alpha component)

\code{is.named.color} checks whether input is a valid named
color
}
\examples{
is.color('turquoise2')
is.hex.color('#8BADF00D')

}
\seealso{
\code{\link{colors}}
}

