% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.colorspace.R
\name{is.colorspace}
\alias{is.colorspace}
\title{Check whether input is a supported colorspace}
\usage{
is.colorspace(x)
}
\arguments{
\item{x}{One or more strings containing a colorspace name}
}
\value{
Boolean indicating whether or not the given input is a valid
colorspace. If more than one are given, ALL must be valid.
}
\description{
\code{is.colorspace} checks whether or not the given input is a
valid colorspace (i.e., one supported by this package)
}
\examples{
is.colorspace('hsv')

}

