% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mix}
\alias{mix}
\title{Add two colors together}
\usage{
mix(color1, color2, weight = 0.5)
}
\arguments{
\item{color1}{The first color}

\item{color2}{The second color}

\item{weight}{The percentage of color2 to be included in the result. The mixed color is obtained by combining an amount 1-alpha of color1 with an amount alpha of color2.}
}
\value{
The mixed color as a hexadecimal string
}
\description{
Add two colors together
}
\examples{
mix('#FFFFFF', '#0000FF', 0.2)

}
\seealso{
mixcolor
}

