% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{saturation}
\alias{blue}
\alias{brightness}
\alias{green}
\alias{hue}
\alias{lightness}
\alias{red}
\alias{saturation}
\title{View a color's properties}
\usage{
saturation(color, space = "HSV")

hue(color, space = "HSV")

brightness(color)

lightness(color)

red(color)

green(color)

blue(color)
}
\arguments{
\item{color}{The color}

\item{space}{The colorspace, HSV (default), HSL, or HLS}
}
\value{
The color's property value
}
\description{
\code{saturation} returns a color's saturation [0,1]

\code{hue} returns a color's hue [0,360)

\code{brightness} returns a color's brightness (value) in HSV space [0,1]

\code{lightness} returns a color's brightness (value) in HSL space [0,1]

\code{red} returns a color's red level [0,1]

\code{green} returns a color's green level [0,1]

\code{red} returns a color's blue level [0,1]
}
\examples{
saturation('orange')

}

