% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb_properties.R
\name{red}
\alias{alpha}
\alias{blue}
\alias{green}
\alias{red}
\title{View a color's RGB components}
\usage{
red(col)

green(col)

blue(col)

alpha(col)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" or
    "\code{#rrggbbaa}"(see \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}
}
\value{
A number indicating the component value [0,1]
}
\description{
\code{red}, \code{green}, \code{blue}, and \code{alpha} return the values for
those components in RGBA space.
}
\examples{
# Get a color's blue-ness
blue("mediumpurple2")
}
\seealso{
\code{\link{adjust_rgb}} to manipulate these components
}

