% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbspace.R
\name{rgbspace}
\alias{adjust_blue}
\alias{adjust_green}
\alias{adjust_red}
\alias{adjust_rgb}
\alias{alpha}
\alias{blue}
\alias{green}
\alias{red}
\alias{rgbspace}
\title{View and adjust the RGB components of a color}
\usage{
red(col)

green(col)

blue(col)

alpha(col)

adjust_rgb(col, Ramount = 0, Gamount = 0, Bamount = 0)

adjust_red(col, amount)

adjust_green(col, amount)

adjust_blue(col, amount)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{Ramount}{Amount to adjust the red component by (default: \code{0})}

\item{Gamount}{Amount to adjust the green component by (default: \code{0})}

\item{Bamount}{Amount to adjust the blue component by (default: \code{0})}

\item{amount}{Amount to adjust the component by (default: \code{0})}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
These functions allow the RGB components of a color---red,
green, blue, and alpha---to be seen and manipulated.

TODO: describe these values and their ranges

\code{red} returns a color's red value [0,1]

\code{green} returns a color's green value [0,1]

\code{blue} returns a color's blue value [0,1]

\code{alpha} returns a color's alpha channel value [0,1]

\code{adjust_rgb} Increases or decreases each component

\code{adjust_red} Increases or decreases the red component of
a color

\code{adjust_green} Increases or decreases the green component
of a color

\code{adjust_blue} Increases or decreases the blue component
of a color
}
\examples{
# Get a color's blueness
blue("mediumpurple2")

# Increase all three components
adjust_rgb('#000000', Ramount = 0.1, Gamount = 0.3, Bamount = 0.1)

# Increase redness by 10\%
adjust_red('#00FFFF', amount = 0.1)

}

