% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation.R
\name{saturation}
\alias{adjust_saturation}
\alias{desaturate}
\alias{grayscale}
\alias{greyscale}
\alias{saturate}
\alias{saturation}
\title{View or adjust a color's saturation}
\usage{
saturation(color)

saturate(color, amount)

desaturate(color, amount)

greyscale(color, amount)
}
\arguments{
\item{color}{One or more colors, either hex or named}

\item{amount}{Amount to adjust saturation by [0,1]}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
\code{saturation} returns the color's saturation [0,1]

\code{saturate} increases saturation by the given amount

\code{desaturate} decreases saturation by the given amount

\code{greyscale} fully desaturates the color
}
\examples{
saturate("orange", 0.1)

}

