% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{shade}
\alias{shade}
\title{Shade a given color, mixing it with black}
\usage{
shade(color, amount)
}
\arguments{
\item{color}{The color to shade}

\item{amount}{The amount of black to add.}
}
\value{
The shaded color as a hexadecimal string
}
\description{
Shade a given color, mixing it with black
}
\examples{
\dontrun{
shade('#FF0000', 0.2)
}

}
\seealso{
mix
}

