% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tint.R
\name{tint}
\alias{tint}
\title{Tint a given color, mixing it with white}
\usage{
tint(color, amount)
}
\arguments{
\item{color}{The color to tint}

\item{amount}{The amount of white to add.}
}
\value{
The tinted color as a hexadecimal string
}
\description{
Tint a given color, mixing it with white
}
\examples{
tint('#FF0000', 0.2)

}
\seealso{
mix
}

