% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparentize.R
\name{transparentize}
\alias{opacify}
\alias{transparentize}
\title{Adjust a color's transparency}
\usage{
transparentize(col, amount, mode = "relative")

opacify(col, amount, mode = "relative")
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{amount}{The amount to increase transparency [-1,1]}

\item{mode}{Whether the adjustments set the value directly}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
\code{transparentize} increases a color's transparency, while \code{opacify}
decreases a color's transparency, making it more opaque.
}

