% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hsv.R
\name{adjust_hsv}
\alias{adjust_hsv}
\title{Adjust color properties in HSV space}
\usage{
adjust_hsv(color, Hamount = 0, Samount = 0, Vamount = 0)
}
\arguments{
\item{color}{One or more colors, either hex or named}

\item{Hamount}{Amount to adjust hue (default: 0)}

\item{Samount}{Amount to adjust saturation (default: 0)}

\item{Vamount}{Amount to adjust value/brightness (default: 0)}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
\code{adjust_hsv} adjusts the properties of colors in HSV space
(hue, saturation, and value/brightness)
}
\note{
Adjusted saturation and brightness are clamped to the range [0,1].

If \code{wraphue} is \code{FALSE}, resulting hue is clamped to the
range [0,360].
}
\examples{
# Decrease saturation of orange by 0.1
adjust_hsv(color = "orange", Samount = -0.1)

lighten('orange', 0.2)

}

