% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hsv.R
\name{adjust_hsv}
\alias{adjust_hsv}
\alias{brighten}
\alias{darken}
\alias{desaturate}
\alias{grayscale}
\alias{greyscale}
\alias{saturate}
\title{Adjust a color's HSV components}
\usage{
adjust_hsv(col, h = NULL, s = NULL, v = NULL, mode = "relative")

saturate(col, amount)

desaturate(col, amount)

greyscale(col, amount)

brighten(col, amount)

darken(col, amount)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{h}{The adjustment in hue [-1,1] (default: no adjustment)}

\item{s}{The adjustment in saturation [-1,1] (default: no adjustment)}

\item{v}{The adjustment in brightness [-1,1] (default: no adjustment)}

\item{mode}{Whether the adjustments set the value directly
("\code{absolute}") or are added to the current value ("\code{relative}",
default).}

\item{amount}{The value of the adjustment}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
These functions adjust a color's hue, saturation, and brightness
(value).

\code{adjust_hsv} adjusts hue, saturation, and brightness either
by replacing the current value (\code{mode="absolute"}) or adding to it
(\code{mode="relative"}, the default).

\code{saturate} and \code{desaturate} increase and decrease
saturation by the given amount, respectively

\code{greyscale} (or \code{grayscale}) fully desaturates the
color

\code{brighten} and \code{darken} increase and decrease
brightness by the given amount, respectively
}
\examples{
# Decrease saturation of orange by 0.1
adjust_hsv(col = "orange", s = -0.1)

}
\seealso{
\code{\link{hue}}, \code{\link{saturation}},
\code{\link{brightness}} to these component values
}

