% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hsv.R
\name{adjust_hsv}
\alias{adjust_hsv}
\alias{complement_HSV}
\alias{darken}
\alias{desaturate_HSV}
\alias{grayscale}
\alias{greyscale_HSV}
\alias{lighten}
\alias{saturate_HSV}
\title{Adjust color properties in HSV space}
\usage{
adjust_hsv(color, Hamount = 0, Samount = 0, Vamount = 0, wraphue = TRUE)

complement_HSV(color, wrap = TRUE)

saturate_HSV(color, amount)

desaturate_HSV(color, amount)

greyscale_HSV(color, space)

lighten(color, amount)

darken(color, amount)
}
\arguments{
\item{color}{One or more colors, either hex or named}

\item{Hamount}{Amount to adjust hue (default: 0)}

\item{Samount}{Amount to adjust saturation (default: 0)}

\item{Vamount}{Amount to adjust value/brightness (default: 0)}

\item{wraphue}{Whether or not the resulting hue should be constrained to
[0,360) (default: TRUE)}

\item{amount}{Amount to adjust the component by (default: 0)}
}
\value{
The adjusted colors as hexadecimal strings
}
\description{
\code{adjust_hsv} adjusts the properties of colors in HSV space (hue, saturation, and value/brightness)

\code{complement_HSV} returns the complement of the color by
adjusting the color's hue by 180 degrees

\code{saturate_HSV} increases saturation by the given amount

\code{desaturate_HSV} decreases saturation by the given amount

\code{greyscale_HSV} fully desaturates the color

\code{lighten} increases the brightness of the color by the
given amount

\code{darken} decreases the brightness of the color by the
given amount
}
\note{
Adjusted saturation and brightness are clamped to the range [0,1].

If \code{wraphue} is \code{FALSE}, resulting hue is clamped to the
range [0,360].
}
\examples{
# Decrease saturation of orange by 0.1
adjust_hsv(color = "orange", Samount = -0.1)

lighten('orange', 0.2)

}

