% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rgb.R
\name{adjust_rgb}
\alias{adjust_blue}
\alias{adjust_green}
\alias{adjust_red}
\alias{adjust_rgb}
\title{Adjust the sRGB components of a color. The resulting components are clamped
to the range [0,1].}
\usage{
adjust_rgb(col, Ramount = 0, Gamount = 0, Bamount = 0)

adjust_red(col, amount)

adjust_green(col, amount)

adjust_blue(col, amount)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{Ramount}{Amount to adjust the red component by ([0,1], default:
\code{0})}

\item{Gamount}{Amount to adjust the green component by ([0,1], default:
\code{0})}

\item{Bamount}{Amount to adjust the blue component by ([0,1], default:
\code{0})}

\item{amount}{Amount to adjust the component by (default: \code{0})}
}
\value{
The adjusted color(s) as a hexadecimal string
}
\description{
Adjust the sRGB components of a color. The resulting components are clamped
to the range [0,1].

\code{adjust_red} Increases or decreases the red component of
a color

\code{adjust_green} Increases or decreases the green component
of a color

\code{adjust_blue} Increases or decreases the blue component
of a color
}
\examples{
adjust_rgb('#000000', Ramount = 0.1, Gamount = 0.3, Bamount = 0.1)
adjust_red('#00FFFF', amount = 0.1)

}

