% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_saturation.R
\name{adjust_saturation}
\alias{adjust_saturation}
\alias{desaturate}
\alias{grayscale}
\alias{greyscale}
\alias{saturate}
\title{Adjust a color's saturation}
\usage{
adjust_saturation(color, amount)

saturate(color, amount)

desaturate(color, amount)

greyscale(color)
}
\arguments{
\item{color}{The color}

\item{amount}{A value between -1 and 1.}
}
\value{
The adjusted color as a hexadecimal string
}
\description{
\code{adjust_saturation} increases the saturation of a color by
the amount given. Resulting saturation is clamped to the range [0,1].

\code{saturate} increases the saturation of the color by the
given amount

\code{desaturate} decreases the saturation of the color by the
given amount

\code{greyscale} fully desaturates the color
}
\note{
Adjustments are made in the HSV colorspace
}
\examples{
desaturate('mediumpurple3', 0.1)

}

