% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brightness.R
\name{brightness}
\alias{brightness}
\title{View or adjust a color's brightness/lightness}
\usage{
brightness(col)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{amount}{Amount to adjust brightness or lightness by [0,1]}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
\code{brightness} returns the color's brightness in HSV
space [0,1]
}
\examples{
brightness("orange")

}

