% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hsv.R
\name{col2hsv}
\alias{col2hsv}
\title{Color to HSV Conversion}
\usage{
col2hsv(color)
}
\arguments{
\item{color}{vector of any of the three kinds of R color specifications TODO (see col2rgb)}
}
\value{
A matrix with a column for each color. The three rows of the matrix
indicate hue, saturation and value and are named "\code{h}", "\code{s}", and
"\code{v}" accordingly.
}
\description{
TODO
}
\examples{
TODO

}

