% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colHLS.R
\name{colHLS}
\alias{colHLS}
\alias{colHSL}
\title{Convert hex or named colors to colorspace HLS objects}
\usage{
colHLS(color)

colHSL(color)
}
\arguments{
\item{color}{One or more colors, either hex or named}
}
\value{
A list of class HLS objects
}
\description{
Convert hex or named colors to colorspace HLS objects
}
\note{
Alpha values are ignored
}
\examples{
colHLS("orange")

}
\seealso{
\code{\link[colorspace]{HLS}}
}

