% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colHSV.R
\name{colHSV}
\alias{colHSV}
\title{Convert hex or named colors to colorspace HSV objects}
\usage{
colHSV(color)
}
\arguments{
\item{color}{One or more colors, either hex or named}
}
\value{
A list of class HSV objects
}
\description{
Convert hex or named colors to colorspace HSV objects
}
\note{
Alpha values are ignored
}
\examples{
colHSV("orange")

}
\seealso{
\code{\link[colorspace]{HSV}}
}

