% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorproperty.R
\name{red}
\alias{alpha}
\alias{blue}
\alias{green}
\alias{red}
\title{Get values of color properties in different color spaces}
\usage{
red(col)

green(col)

blue(col)

alpha(col)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{space}{The color space}

\item{property}{The property value to get. For example, when \code{space} is
\code{HSV}, \code{property} can be "H", "S", or "V".}
}
\value{
Numeric property values
}
\description{
Get values of color properties in different color spaces
}
\examples{
#TODO
}

