% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colRGB.R
\name{colRGB}
\alias{colRGB}
\title{Convert a hex or named color to a colorspace RGB object}
\usage{
colRGB(color)
}
\arguments{
\item{color}{The color}
}
\value{
An object of class “RGB” which inherits from class “color”.
}
\description{
Convert a hex or named color to a colorspace RGB object
}
\note{
Alpha values are ignored
}
\seealso{
\code{\link{RGB}}
}

