% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colRGB.R
\name{colRGB}
\alias{colRGB}
\title{Convert hex or named colors to colorspace RGB objects}
\usage{
colRGB(color)
}
\arguments{
\item{color}{One or more colors, either hex or named}
}
\value{
A list of class sRGB objects
}
\description{
Convert hex or named colors to colorspace RGB objects
}
\note{
Alpha values are ignored
}
\examples{
colRGB("orange")

}
\seealso{
\code{\link[colorspace]{RGB}}
}

