% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement}
\alias{complement}
\title{Get a color's complement}
\usage{
complement(col)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
\code{complement} returns a color's complement, which is the hue adjusted by
180 degrees
}
\examples{
# Get orange's complement
complement("orange")

}

