% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex}
\alias{hex}
\title{Convert colors to hexadecimal strings}
\usage{
hex(color)
}
\arguments{
\item{color}{One or more colors, either hex or named}
}
\value{
A vector of colors as hex strings
}
\description{
Convert colors to hexadecimal strings
}
\examples{
hex(c('mediumpurple3', 'orange'))

}

