% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex}
\alias{hex}
\title{Get a hex string representation of colors}
\usage{
hex(col, alpha)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" or
    "\code{#rrggbbaa}"(see \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{alpha}{Whether or not to include alpha channel values in the result.}
}
\value{
The color(s) as hexadecimal strings
}
\description{
Get a hex string representation of colors
}
\examples{
hex("orange", alpha = FALSE)

}

