% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv2col.R
\name{hsv2col}
\alias{hsv2col}
\title{HSV to R color conversion}
\usage{
hsv2col(color)
}
\arguments{
\item{color}{TODO}
}
\value{
A vector of colors corresponding to the given values in HSV space.
}
\description{
\code{hsv2col} transforms colors from HSV space into   TODO
}
\examples{
TODO

}

