% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv2rgb.R
\name{hsv2rgb}
\alias{hsv2rgb}
\title{HSV to RGB conversion}
\usage{
hsv2rgb(color, alpha = FALSE)
}
\arguments{
\item{color}{TODO copy from col2rgb}
}
\value{
An integer matrix with three or four (for \code{alpha = TRUE}) rows
and number of columns the length of col. If col has names these are used as
the column names of the return value.
}
\description{
TODO

\code{hsv2rgb} transforms colors from HSV space into RGB space
}
\examples{
TODO

}

