% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv2rgb.R
\name{hsv2rgb}
\alias{hsv2rgb}
\title{HSV to RGB conversion}
\usage{
hsv2rgb(col)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}
}
\value{
An integer matrix with three rows
and number of columns the length of col. If col has names these are used as
the column names of the return value.
}
\description{
TODO

\code{hsv2rgb} transforms colors from HSV space into RGB space
}
\examples{
TODO

}

