% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hue.R
\name{hue}
\alias{adjust_hue}
\alias{complement}
\alias{hue}
\title{View or adjust a color's hue}
\usage{
hue(color)

adjust_hue(color, amount)

complement(color)
}
\arguments{
\item{color}{One or more colors, either hex or named}

\item{amount}{Amount to adjust hue by [0,1]}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
\code{hue} returns the color's hue [0,360)

\code{adjust_hue} Increases or decreases the color's hue

\code{complement} returns a color's complement, which is the hue
adjusted by 180 degrees
}
\examples{
hue("orange")

complement("orange")

}

