% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert}
\alias{invert}
\title{Invert a color}
\usage{
invert(col)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}
}
\value{
The inverted color as a hexadecimal string
}
\description{
\code{invert} inverts each channel of a color in RGB space
}
\note{
Alpha channel values are be stripped from \code{col}
}
\examples{
invert('#FF00FF')

}

