% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mix}
\alias{mix}
\alias{shade}
\alias{tint}
\title{Color mixing}
\usage{
mix(col1, col2, amount = 0.5)

shade(col, amount)

tint(col, amount)
}
\arguments{
\item{col1}{One or more colors, either hex or named}

\item{col2}{One or more colors, either hex or named}

\item{amount}{The amount of \code{col2} to add. For \code{shade} and \code{tint},
this is the amount of black and white, respectively, to add.}

\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}
}
\value{
The mixed colors as hexadecimal strings
}
\description{
\code{mix} mixes two colors together

\code{shade} shades a given color, mixing it with black

\code{tint} tints a given color, mixing it with white
}
\examples{
mix('#FFFFFF', '#0000FF', 0.2)

}

