% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mix}
\alias{mix}
\alias{shade}
\alias{tint}
\title{Color mixing}
\usage{
mix(color1, color2, weight = 0.5)

shade(color, amount)

tint(color, amount)
}
\arguments{
\item{color1}{The first color}

\item{color2}{The second color}

\item{weight}{The percentage of color2 to be included in the result. The mixed color is obtained by combining an amount 1-alpha of color1 with an amount alpha of color2.}
}
\value{
The mixed color as a hexadecimal string
}
\description{
\code{mix} mixes two colors together

\code{shade} shades a given color, mixing it with black

\code{tint} tints a given color, mixing it with white
}
\examples{
mix('#FFFFFF', '#0000FF', 0.2)

}
\seealso{
mixcolor
}

