% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mix}
\alias{mix}
\alias{shade}
\alias{tint}
\title{Color mixing}
\usage{
mix(color1, color2, amount = 0.5)

shade(color, amount)

tint(color, amount)
}
\arguments{
\item{color1}{One or more colors, either hex or named}

\item{color2}{One or more colors, either hex or named}

\item{amount}{The amount of color2 to add. For \code{shade} and \code{tint}, this is the amount of black and white, respectively, to add.}

\item{color}{One or more colors, either hex or named}
}
\value{
The mixed colors as hexadecimal strings
}
\description{
\code{mix} mixes two colors together

\code{shade} shades a given color, mixing it with black

\code{tint} tints a given color, mixing it with white
}
\examples{
mix('#FFFFFF', '#0000FF', 0.2)

}
\seealso{
\code{\link[colorspace]{mixcolor}}
}

