% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{brightness}
\alias{brightness}
\alias{lightness}
\title{View a color's properties}
\usage{
brightness(color)

lightness(color)
}
\arguments{
\item{color}{The color}

\item{space}{The colorspace, HSV (default), HSL, or HLS}
}
\value{
The color's property value
}
\description{
\code{saturation} returns a color's saturation [0,1]

\code{brightness} returns a color's brightness (value) in HSV space [0,1]

\code{lightness} returns a color's brightness (value) in HSL space [0,1]
}
\examples{
hue('orange')

}

