% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2hex.R
\name{rgb2hex}
\alias{rgb2hex}
\title{RGB to Hex Color Conversion}
\usage{
rgb2hex(col, alpha = FALSE)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{alpha}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO

}

