% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2hex.R
\name{rgb2hex}
\alias{rgb2hex}
\title{RGB to Hex Color Conversion}
\usage{
rgb2hex(color, alpha = FALSE)
}
\arguments{
\item{color}{TODO}

\item{alpha}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO

}

