% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation.R
\name{saturation}
\alias{adjust_saturation}
\alias{desaturate}
\alias{grayscale}
\alias{greyscale}
\alias{saturate}
\alias{saturation}
\title{View or adjust a color's saturation}
\usage{
saturation(col)

saturate(col, amount)

desaturate(col, amount)

greyscale(col, amount)
}
\arguments{
\item{col}{Vector containing any of the three kinds of \R color
specifications: \enumerate{
    \item a color name (see {\link[grDevices]{colors}})
    \item a hexadecimal string of the form "\code{#rrggbb}" (see
    \code{\link[grDevices]{rgb}})
    \item a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}
}}

\item{amount}{Amount to adjust saturation by [0,1]}
}
\value{
The adjusted color(s) as hexadecimal strings
}
\description{
\code{saturation} returns the color's saturation [0,1]

\code{saturate} increases saturation by the given amount

\code{desaturate} decreases saturation by the given amount

\code{greyscale} fully desaturates the color
}
\examples{
saturate("orange", 0.1)

}

