% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation.R
\name{saturation}
\alias{desaturate}
\alias{grayscale}
\alias{greyscale}
\alias{saturate}
\alias{saturation}
\title{View or adjust a color's saturation}
\usage{
saturation(color, space = "HSV")

saturate(color, amount, space = "HSV")

desaturate(color, amount, space = "HSV")

greyscale(color, space = "HSV")
}
\arguments{
\item{color}{One or more colors, either hex or named}

\item{space}{The colorspace to use, "\code{HSV}" (default) or "\code{HSL}"}

\item{amount}{Amount to adjust saturation by [0,1]}
}
\value{
The adjusted colors as hexadecimal strings
}
\description{
\code{saturation} returns the color's saturation [0,1]

\code{saturate} increases saturation by the given amount

\code{desaturate} decreases saturation by the given amount

\code{greyscale} fully desaturates the color
}
\examples{
saturate("orange", 0.1)

}

